# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "CS API",
        "description": "Convenience Select",
        "version": "1.0.0"
    },
    "host": "localhost:9000",
    "schemes": [
        "http",
        "https"
    ],
    "basePath": "/api/v1",
    "produces": [
        "application/json"
    ],
    "paths": {
        "/auth/local/signup": {
            "post": {
                "operationId": "postCredential",
                "summary": "Sign up user with email and password.",
                "description": "Add credential to user <br /> Public API. Not authenticated users can consume it. <br /> Only applies to User with no createdBy attribute. <br /> Only applies to User with no existing credentials.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "User model",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postAuthCredential"
                        }
                    }
                ],
                "tags": [
                    "Auth"
                ],
                "responses": {
                    "200": {
                        "description": "Successful user credential added",
                        "schema": {
                            "$ref": "#/definitions/postAuthResponse"
                        }
                    },
                    "400": {
                        "description": "Validation field error",
                        "schema": {
                            "$ref": "ValidationError"
                        }
                    },
                    "403": {
                        "description": "User permission error",
                        "schema": {
                            "$ref": "AuthCredentialPermission"
                        }
                    },
                    "404": {
                        "description": "User does not exists error",
                        "schema": {
                            "$ref": "AuthCredentialNotExists"
                        }
                    },
                    "409": {
                        "description": "User credentials conflict error",
                        "schema": {
                            "$ref": "AuthCredentialExists"
                        }
                    }
                }
            }
        },
        "/auth/local/login": {
            "post": {
                "operationId": "postCredentialLogin",
                "summary": "Login user with email and password.",
                "description": "Login user into the system",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "User login credentials",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postAuthLogin"
                        }
                    }
                ],
                "tags": [
                    "Auth"
                ],
                "responses": {
                    "200": {
                        "description": "Successful user credential added",
                        "schema": {
                            "$ref": "#/definitions/postAuthResponse"
                        }
                    },
                    "400": {
                        "description": "Validation field error",
                        "schema": {
                            "$ref": "ValidationError"
                        }
                    },
                    "403": {
                        "description": "User permission error",
                        "schema": {
                            "$ref": "AuthCredentialPermission"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "postAuthCredential": {
            "required": [
                "userId",
                "email",
                "password"
            ],
            "properties": {
                "userId": {
                    "type": "string",
                    "description": "AlfaNum User unique id"
                },
                "email": {
                    "type": "string",
                    "description": "User email to sign up"
                },
                "password": {
                    "type": "string",
                    "description": "User Passwocrd"
                },
                "rememberMe": {
                    "type": "boolean",
                    "description": "Remember me login expire unlimited"
                }
            }
        },
        "postAuthLogin": {
            "required": [
                "email",
                "password",
                "rememberMe"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "description": "User email to login"
                },
                "password": {
                    "type": "string",
                    "description": "User password"
                },
                "rememberMe": {
                    "type": "boolean",
                    "description": "Remember me login expire unlimited"
                }
            }
        },
...
        "postAuthResponse": {
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string",
                    "description": "AlfaNum token unique id"
                }
            }
        },
...
    }
}